move ccxt and dydx to their own modules
