+++ pdr_backend/exchange/test/test_fetch_ohlcv_ccxt.py
@@ -1,7 +1,7 @@
 from enforce_typing import enforce_types
 import pytest
 
-from pdr_backend.exchange.fetch_ohlcv import safe_fetch_ohlcv_ccxt
+from pdr_backend.exchange.fetch_ohlcv_ccxt import safe_fetch_ohlcv_ccxt
 from pdr_backend.util.time_types import UnixTimeMs
